Removing file.rs, adding some docs, exporting MultipartRequestBuilder.
